feature: Trend Calculation
This commit will add trend calculation to the seeding status module. This will help the bot "understand" whether player count is truly going up or down, and will indicate as such.
Trend calculation works by using datapoints in the database which are analyzed prior to sending seed messages.
Overall this is a feature improvement of the previous logic.